Initial revision.
This still needs a bit of work to have more comprehensive coverage of
the various cases. This is NOT a good example of how to use the scripts
for formatting, but instead as a way of testing changes to the
underlying script. See the README for details on how to use the
formatter…