Adding new features.
Most of the support for variables and links, inline, reference, aliases
and explicit are in. Still need quite a bit of cleanup and bug fixes
for edge cases. Need to remove old code. Still more to go, but will
have to take some time off to work on other projects.