Added push() method
Needed a way to push an open file onto the filestack, so callers have a way of setting stdin to an open file handled instead of always requiring a filename. Also, needed to remember when we have read from sys.stdin, so that we can continue reading from it once a file is pushed onto the stack.