Split the open out of parse and fix bugs
Can't have parse() be responsible for opening the file, in case callers want to be able to open their own input and just set the file handle. Fixed a bug where I didn't pass in the args when the av_parse_file() method was invoked. Also fixed a bug where I would stop reading if a class was put on a line by itself.