Added ability to reprocess a line that didn't match any rules
This happens when the only thing on a line is a [variable] or {:.aclass}[variable]. Before, the line would expand according to the variable contents, and get wrapped in a <div class="extras">, which wasn't ideal. With this change, you can now have a variable that expands into a new rule, that when reprocessed, will match as expected.