Fixed a number of bugs discovered while writing the unit test code
Made namespace a property using a property decorator.  Made it so BasicNamespace() throws an exception if you call it with a dictionary. Don't pass the namespace down, since you know the namespace inside each object. Use it. Fix bug where I didn't always use an explicity namespace identifier when using self. and calling markdown(). Pass in an API during NS init so we can pass the NS object over to markdown. Fix a number of bugs where I was using the wrong variable_name when calling the namespace methods.