Integrated new variable and markdown classes
Commented out all of the old stuff, so I'd have a revision that had both of them in it, for comparison in the future. Next round I need to to remove all the old code, and also add the @revision and @contact defaults to built-ins, so I can remove unused keywords.