Added public/private keys only and fixed bugs
When doing inline help, sometimes you just need the keys. Actually raise a LogicError() properly when needed. When marking down values in unescapeString, make sure you have a string. Changed behavior of @code.varname. When no attributes are specified, it now default to .run, instead of printing the attributes, which was kind of dumb...