Cleaned up and documented the @code namespace ...
Also fixed how parameters work; they are now forced into scope while a variable is in the getValue() state, so that if, while expanding, there are other references to attrs within it which were passed as parameters, you can now see them. Duh. Everything is restored on the way out, preservering the "non-stickiness" behaviour of @code attributes.