Various enhancements (see details)
Added encodeURL for encoding email addresses and links

This is done on the fly, when the HREF attribute is requested from any LINK namespace variable. May need to look at making it more generic in the future.

Added wrapper for compile() and do exception handling so it's not so easy to crash avscript.py.

Improved printing of dictionaries in debug output, and also styled many of the messages so it's easier to interpret the output.