Added several new features
I think _raw is going to be removed, but I'm checking it in to track it.

Added getAttribute in VariableStore, and then built out the supporting methods in the Namespaces class.

Suspend/Resume debug messages while code is being eval()'ed or exec()'ed. This prevents debug messages from being mixed in with the output from eval() and exec().